Second code revision

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -9,16 +9,16 @@ public class StatsService {
 //    Above average     - количество месяцев, в которых продажи были выше среднего (см. п.2).
 
 
-    public int sumSales(long[] arrs) {  // сумма всех продаж за все месяцы.
-        int sumSales = 0;
+    public long sumSales(long[] arrs) {  // сумма всех продаж за все месяцы.
+        long sumSales = 0;
         for (int i = 0; i < arrs.length; i++) {
-            sumSales = (int) (sumSales + arrs[i]);
+            sumSales =  (sumSales + arrs[i]);
         }
         return sumSales;
     }
 
-    public int averageSale(long[] arrs) {   // средняя сумма продаж
-        int averageSale = sumSales(arrs) / (arrs.length);
+    public long averageSale(long[] arrs) {   // средняя сумма продаж
+        long averageSale = sumSales(arrs) / (arrs.length);
         return averageSale;
     }
 
@@ -46,7 +46,7 @@ public class StatsService {
 
     public int belowTheAverage(long[] arrs) {  // количество месяцев, в которых продажи были ниже среднего
         int monthsWithMinimalSales = 0;   // Months with minimal sales  Месяцы с минимальными продажами
-        int average = averageSale(arrs);
+        long average = averageSale(arrs);
         for (int i = 0; i < arrs.length; i++) {
             if (arrs[i] < average)
                 monthsWithMinimalSales++;
@@ -57,7 +57,7 @@ public class StatsService {
 
     public int aboveAverage(long[] arrs) {  // количество месяцев, в которых продажи были выше среднего
         int monthsWithMaxSales = 0;   //High sales months  Месяцы с высокими продажами
-        int average = averageSale(arrs);
+        long average = averageSale(arrs);
         for (int i = 0; i < arrs.length; i++) {
             if (arrs[i] > average)
                 monthsWithMaxSales++;
